Display leave button always

+++ scripts/sea-battle.js
@@ -430,6 +430,8 @@ new Vue({
         },
 
         quitGame() {
+            if (!confirm('Вы действительно хотите покинуть игру?') === true) return;
+
             localStorage.setItem('seaBattle', JSON.stringify({user: this.user}));
             this.isGameInProgress = false;
             this.opponent.name = null;
